requirement.txt
anyio==4.9.0
certifi==2025.7.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.1.1
python-telegram-bot==22.3
sniffio==1.3.1
requests>=2.31.0,<3.0.0
